move commands into separate dir
