+++ webdriver/src/commands/flyweight.ts
@@ -1,0 +1,21 @@
+import type { IncomingMessage } from "http";
+
+export const jsonHeaders = new Headers([["Content-Type", "application/json"]]);
+
+export function getJsonFromRequestBody(req: IncomingMessage): Promise<any> {
+	return new Promise(function (resolve, reject) {
+		let data: Uint8Array[] = [];
+		req.addListener("data", function (chunk) {
+			data.push(chunk);
+		});
+		req.addListener("end", function () {
+			let actionStr = Buffer.concat(data).toString();
+			let action = JSON.parse(actionStr);
+
+			resolve(action);
+		});
+		req.addListener("error", function (err: Error) {
+			reject(err);
+		});
+	});
+}
