+++ webdriver/src/commands/sessions.ts
@@ -1,0 +1,63 @@
+import type { IncomingMessage, ServerResponse } from "http";
+
+// BELOW ARE ACTIONS FROM TESTS THEMSELVES
+import type { WebdriverParams } from "../config.js";
+import type { EventBusInterface } from "../eventbus.js";
+
+import { jsonHeaders } from "./flyweight.js";
+
+// NEW SESSION
+export async function newSession(
+	params: WebdriverParams,
+	signal: AbortSignal,
+): Promise<string> {
+	let { url, capabilities } = params;
+
+	let res = await fetch(new URL("/session", url), {
+		method: "POST",
+		headers: jsonHeaders,
+		body: JSON.stringify({ capabilities: capabilities ?? {} }),
+		signal,
+	});
+	if (200 !== res.status) {
+		let cause = await res.text();
+		throw new Error("Failed to create a session", { cause });
+	}
+
+	let json = await res.json();
+	let { sessionId } = json?.value;
+	if (typeof sessionId !== "string") throw new Error("session is not a string");
+
+	return sessionId;
+}
+
+// DELETE SESSION
+export async function deleteSession(
+	params: WebdriverParams,
+	signal: AbortSignal | undefined,
+	eventbus: EventBusInterface,
+	sessionId: string | undefined,
+): Promise<void> {
+	let { url, jackrabbitId } = params;
+	try {
+		let delReqest = await fetch(new URL(`/session/${sessionId}`, url), {
+			method: "DELETE",
+			headers: jsonHeaders,
+			body: null,
+			signal: signal,
+		});
+		if (200 !== delReqest.status) {
+			let cause = await delReqest.json();
+			throw new Error("delete-cookie request failed", { cause });
+		}
+	} catch (e) {
+		eventbus.dispatchAction({
+			type: "log",
+			jackrabbitId,
+			loggerAction: {
+				type: "session_error",
+				error: e?.toString() ?? "failed to delete browser session error",
+			},
+		});
+	}
+}
