+++ webdriver/src/webdriver.ts
@@ -3,7 +3,8 @@ import type { ConfigInterface, WebdriverParams } from "./config.js";
 import type { EventBusInterface } from "./eventbus.js";
 
 import { exec } from "child_process";
-import { deleteSession, newSession, addCookie, go } from "./commands.js";
+import { addCookie, go } from "./commands.js";
+import { deleteSession, newSession } from "./commands/mod.js";
 import { untilWebdriverReady } from "./operations.js";
 
 export class WebDrivers {
